changes for removing error

+++ ejbModule/service/CommitteeMemberService.java
@@ -8,14 +8,14 @@ import exception.YearbookException;
 public interface CommitteeMemberService {
 
 
-	public boolean updateMember(int memberId, String fName, String lName,
+	public boolean updateMember(long memberId, String fName, String lName,
 			String designation, int deptId,  String photoUrl) throws YearbookException;
 	
 	
 	public boolean addMember(String fName, String lName,
 			String designation, int deptId,  String photoUrl) throws YearbookException;
 
-	public boolean deleteMember(int memberId) throws YearbookException;
+	public boolean deleteMember(long memberId) throws YearbookException;
 	
 	public List<CommitteeMember> getAllCommitteeMembers(int deptId) throws YearbookException;
